feature-stats/src/main/java/com/arsvechkarev/stats/presentation/StatsViewModel.kt
package com.arsvechkarev.stats.presentation

import com.arsvechkarev.stats.domain.StatsUseCase
import core.BaseScreenState
import core.Failure
import core.Loading
import core.NetworkAvailabilityNotifier
import core.NetworkListener
import core.Schedulers

class StatsViewModel(
  private val statsUseCase: StatsUseCase,
  private val networkAvailabilityNotifier: NetworkAvailabilityNotifier,
  private val schedulers: Schedulers
) : base.RxViewModel(), NetworkListener {
  
  init {
    networkAvailabilityNotifier.registerListener(this)
  }
  
  override fun onNetworkAvailable() {
    if (_state.value is Failure) {
      schedulers.mainThread().scheduleDirect(::startLoadingData)
    }
  }
  
  fun startLoadingData() {
    rxCall {
      statsUseCase.getWorldCasesInfo()
          .subscribeOn(schedulers.io())
          .map<BaseScreenState>(::LoadedMainStatistics)
          .onErrorReturn(::Failure)
          .startWith(Loading)
          .observeOn(schedulers.mainThread())
          .smartSubscribe(_state::setValue)
    }
  }
  
  override fun onCleared() {
    networkAvailabilityNotifier.unregisterListener(this)
  }
}